refactor reset password to page pattern
